requirements.txt
typer>=0.9.0
PyYAML==6.0
